Edit rubric items works!
Added functionality to be able to edit rubric items that were entered on
the addBlankRubric and editRubric views. Also cleaned up the
look/display of edit rubrics.
Still needs front end work to make it look better.
Note: next I will figure out how to make the fields be required for the
editing. If the itemCategory is left blank, app crashes.